scope was made generic and moved into its own file
