+++ src/lib.rs
@@ -2,3 +2,4 @@ mod lexer;
 mod lson;
 mod parse;
 mod runtime;
+mod scope;
